backend/requirements-dev.txt
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0
autopep8==2.0.2
coverage==7.3.0
flasgger==0.9.7.1
python-dotenv==1.0.0
